+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/language/GitlabYamlFileType.kt
@@ -1,0 +1,24 @@
+package com.github.blarc.gitlab.template.lint.plugin.language
+
+import com.github.blarc.gitlab.template.lint.plugin.ui.Icons
+import com.intellij.openapi.fileTypes.LanguageFileType
+import org.jetbrains.yaml.YAMLLanguage
+import javax.swing.Icon
+
+class GitlabYamlFileType : LanguageFileType(YAMLLanguage.INSTANCE, true) {
+    override fun getName(): String {
+        return "Gitlab Yaml"
+    }
+
+    override fun getDescription(): String {
+        return "Gitlab Yaml"
+    }
+
+    override fun getDefaultExtension(): String {
+        return "gitlab-ci.yml"
+    }
+
+    override fun getIcon(): Icon {
+        return Icons.GITLAB_LOGO
+    }
+}