+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/ui/Icons.kt
@@ -4,4 +4,5 @@ import com.intellij.openapi.util.IconLoader
 
 object Icons {
     val GITLAB_LINT = IconLoader.getIcon("/icons/gitlabLint.svg", javaClass)
+    val GITLAB_LOGO = IconLoader.getIcon("/icons/gitlabLogo.svg", javaClass)
 }